configured profiles/mappers for all dbo's, moved base query classes to Base/
